services/deployment/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
boto3==1.34.0
python-terraform==0.10.1
python-dotenv==1.0.0
asyncpg==0.29.0
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
pyyaml==6.0.1
jinja2==3.1.2